[CP-77] Added support for getting usage statistics from an organization by spaces.
Made the usage service, this because the space usage does not return one space's usage, but the organizations usage queried by space. Thus it is unnecessary to create the getUsage method in the space and in the organizations services. I implemented these methods in the usage service and the corresponding unit tests in the _test file.