[CP-31] Added support for creating/uploading resources.
New Create method added inside the Resources service. In contentful.go, I edited the Do method, which (amongst other things) decodes the response. The response returned by the upload API cannot be decoded, so the code now skips the decode if the method-caller is the Create method of the Resource service. Also added the unit test for the Create method.